pyproject.toml
[tool.poetry]
name = "projeto"
version = "0.1.0"
description = ""
authors = ["Leandro F. <leandro@licitegov.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
django-oauth-toolkit = "^2.2.0"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
